Proxy requests to `pixlet serve` through the Flask app
This removes the need to expose the ports used by Pixlet.

Change the way the Flask app interact with Pixlet: instead of pointing
the client to the pixlet instance's port, let the Flask app forward
requests.

With this change, the app only exposes a single port which improves
scalability (no need to adjust port configs after adding a user) allows
TLS configuration with a given certificate only using gunicorn (example
coming in a future PR).

The pixlet processes are still allocated static port numbers, which
could be improved independently.