Remove websocket proxy for pixlet
It didn't really work because there was no websocket server running. The websocket is used by Pixlet to watch for changes in the applet source files which isn't very useful within tronbyt-server. Apps could change within a container when refreshing a repository, but in that case, the configapp page should just be refreshed.

I looked into flask-sockets, but this is deprecated. The replacement flask-sock doesn't have a Debian package, so I don't think this is really worth it.