Redesign app configuration page (#167) (#168)
Instead of running `pixlet serve` on per-user ports, use libpixlet's
`get_schema` to retrieve the schema for the given app and render a form
to configure the app. This is faster because we don't have to wait for
an external process to start and simplifies the code a lot. We can also
apply more consistent styling to the form.

It is not feature complete yet:
- `locationbased` settings are not handled yet because they need another
  call into Starlark
- settings of type `location` should get a new picker which can then also
  be used on the device page to set the device location

For this reason, the new flow can be activated for testing by setting
the environment variable `USE_PIXLET_SCHEMA` to `1`.

Once this is done, we can delete more code:
- the routes which forward request to `pixlet serve`
- the assignment of ports to users
- the `pixlet` binary itself
- the temporary config files
- the iframe hosting `pixlet serve`

Co-authored-by: Ingmar Stein <IngmarStein@users.noreply.github.com>