Separate dynamic from static data
In order to distribute tronbyt-server as a package, the directory layout
needs to be changed. The package will contain static data like code and
templates and live in a read-only location, so we can no longer write
data like rendered images, the system app repo, app metadata, and app
previews to this location.

This change moves the dynamic data to a consolidated location. In a
container, this is mounted as /app/data, but it can be configured for
bare-metal installations. This move causes a small one-time
discontinuity, but no data loss:
- the system apps repo will be cloned one again
- the app previews will be broken until each app is rendered again

Before:
- /app/system-apps: dynamic data (discardable)
- /app/system-apps.json: dynamic data (discardable)
- /app/tronbyt_server/webp: dynamic data (discardable)
- /app/tronbyt_server/static/apps: dynamic data (discardable)
- /app/users: user-generated data (should be preserved)
- /app (everything else): static data

After:
- /app/data: dynamic data (discardable)
- /app/users: user-generated data (should be preserved)
- /app (everything else): static data

The dynamic data is discardable because it will be regenerated, but it's
still useful to preserve to speed up container restarts, for example.

This change also reduces the recommended number of Docker volumes from
three (apps, users, webp) to two (users, data).